Move CASTOR L1 emulator into own module. Restore BuildFile and SealModule to last reversion.
