+++ SimCalorimetry/CastorSim/plugins/SealModule.cc
@@ -2,11 +2,9 @@
 #include "SimCalorimetry/CastorSim/plugins/CastorDigiProducer.h"
 #include "SimCalorimetry/CastorSim/plugins/CastorHitAnalyzer.h"
 #include "SimCalorimetry/CastorSim/plugins/CastorDigiAnalyzer.h"
-#include "SimCalorimetry/CastorSim/plugins/CastorTTRecord.h"
 #include "SimGeneral/MixingModule/interface/DigiAccumulatorMixModFactory.h"
 
 DEFINE_DIGI_ACCUMULATOR(CastorDigiProducer);
 DEFINE_FWK_MODULE(CastorHitAnalyzer);
 DEFINE_FWK_MODULE(CastorDigiAnalyzer);
-DEFINE_FWK_MODULE(CastorTTRecord);
