+++ SimCalorimetry/CastorTechTrigProducer/src/SealModule.cc
@@ -1,0 +1,5 @@
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+#include "FWCore/Framework/interface/MakerMacros.h"
+#include "SimCalorimetry/CastorTechTrigProducer/src/CastorTTRecord.h"
+
+DEFINE_FWK_MODULE(CastorTTRecord);
